NTBS-195 remove unnecessary v-class binding, add SearchParameters class and move some filtering methods to service
