+++ ntbs-service/Models/Notification.cs
@@ -23,7 +23,9 @@ namespace ntbs_service.Models
         
         [Display(Name = "Notification Id")]
         public int NotificationId { get; set; }
+        [MaxLength(50)]
         public string ETSID { get; set; }
+        [MaxLength(50)]
         public string LTBRID { get; set; }
         public DateTime CreationDate { get; set; }
         public DateTime? SubmissionDate { get; set; }
