+++ ntbs-service/Pages/Search/Index.cshtml
@@ -12,25 +12,24 @@
 <div>
     <form method="get">
         <div asp-validation-summary="ModelOnly"></div>
-        <div class="enter-search-details-tip">
+        <span class="enter-search-details-tip">
             To search enter at least one of the following fields
-        </div>
+        </span>
         <nhs-grid-row>
             <nhs-grid-column grid-column-width="OneQuarter">
                 @{
-                    var IdFilterError = !Model.IsValid("IdFilter");
+                    var IdFilterError = !Model.IsValid("SearchParameters.IdFilter");
                     var IdFilterGroupState = IdFilterError ? Error : Standard;
                 }
-                <nhs-form-group nhs-form-group-type=@IdFilterGroupState v-bind:class="{ 'nhsuk-form-group--error': hasError}">
+                <nhs-form-group nhs-form-group-type=@IdFilterGroupState>
                     <label nhs-label-type="Standard"> ID number </label>
                     <span class="nhsuk-hint " id="example-hint">
                         Search by NTBS, ETS, LTBR or NHS number
                     </span>
                     <span ref="errorField" nhs-span-type="ErrorMessage" id="id-error"
-                        asp-validation-for="IdFilter"></span>
+                        asp-validation-for="SearchParameters.IdFilter"></span>
                     <input v-on:blur="validate" is-error-input=@IdFilterError nhs-input-type="Standard"
-                        asp-for="IdFilter" classes="nhsuk-input--width-10"
-                        v-bind:class="{ 'nhsuk-input--error': hasError}" />
+                        asp-for="SearchParameters.IdFilter" classes="nhsuk-input--width-10" />
                 </nhs-form-group>
             </nhs-grid-column>
         </nhs-grid-row>
@@ -49,8 +48,11 @@
             </a>
         </div>
         <div>
-            @{ var resultCount = Model.SearchResults?.NumberOfResults ?? 0; }
-            <p> @resultCount Results found. </p>
+            @{ 
+                var resultCount = Model.SearchResults?.NumberOfResults ?? 0; 
+                var resultsFoundString = resultCount.ToString() + " Result" + (resultCount == 1 ? "" : "s") + " found.";
+            }
+            <p> @resultsFoundString </p>
             <p> To filter results, add more search terms and search again. </p>
         </div>
         @if(Model.SearchResults != null) 
@@ -59,6 +61,7 @@
                 @foreach(var result in @Model.SearchResults) 
                 {
                     <li>
+                        <!-- Each search hit is displayed using a notification banner -->
                         <partial name="../Shared/_NotificationBanner.cshtml" model="result"/>
                     </li>
                 }
