+++ ntbs-service/Pages/Shared/_NoForm_Layout.cshtml
@@ -2,7 +2,7 @@
     Layout = "./_Layout.cshtml";
 }
 
-<nhs-grid-row style="position: relative">
+<nhs-grid-row classes="body-and-aside-container">
     <nhs-grid-column grid-column-width="Full">
         @RenderBody()
     </nhs-grid-column>
