+++ ntbs-service/Services/NotificationService.cs
@@ -28,6 +28,8 @@ namespace ntbs_service.Services
         Task UpdatePatientTBHistoryAsync(Notification notification, PatientTBHistory history);
         Task UpdateSocialRiskFactorsAsync(Notification notification, SocialRiskFactors riskFactors);
         Task<Notification> CreateLinkedNotificationAsync(Notification notification);
+        IQueryable<Notification> FilterById(IQueryable<Notification> IQ, string IdFilter);
+        IQueryable<Notification> OrderQueryableByNotificationDate(IQueryable<Notification> query);
     }
 
     public class NotificationService : INotificationService
@@ -226,6 +228,16 @@ namespace ntbs_service.Services
         {
             context.AddAuditCustomField(CustomFields.AuditDetails, auditType);
             await context.SaveChangesAsync();
+        }
+
+        public IQueryable<Notification> FilterById(IQueryable<Notification> IQ, string IdFilter) {
+            return IQ.Where(s => s.NotificationId.Equals(Int32.Parse(IdFilter)) 
+                    || s.ETSID.Equals(IdFilter) || s.LTBRID.Equals(IdFilter) || s.PatientDetails.NhsNumber.Equals(IdFilter));
+        }
+
+        public IQueryable<Notification> OrderQueryableByNotificationDate(IQueryable<Notification> query) {
+            return query.OrderByDescending(n => n.CreationDate)
+                .OrderByDescending(n => n.SubmissionDate);
         }
     }
 }
