+++ ntbs-service/wwwroot/css/notification.scss
@@ -14,6 +14,10 @@
     }
 }
 
+.body-and-aside-container {
+    position: relative;
+}
+
 .search-notifications-button {
     font-size: medium;
 }
